move inUse exception to correct package
